pyproject.toml
[project]
name = "auth-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "brevo-python>=4.0.6",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
